// client/stocklinx app/src/redux/department/constant.ts
export enum departmentConst {
  FETCH_DEPARTMENTS_REQUEST = "FETCH_DEPARTMENTS_REQUEST",
  FETCH_DEPARTMENTS_SUCCESS = "FETCH_DEPARTMENTS_SUCCESS",
  FETCH_DEPARTMENTS_FAILURE = "FETCH_DEPARTMENTS_FAILURE",
  FETCH_DEPARTMENT_REQUEST = "FETCH_DEPARTMENT_REQUEST",
  FETCH_DEPARTMENT_SUCCESS = "FETCH_DEPARTMENT_SUCCESS",
  FETCH_DEPARTMENT_FAILURE = "FETCH_DEPARTMENT_FAILURE",
  CREATE_DEPARTMENT_REQUEST = "CREATE_DEPARTMENT_REQUEST",
  CREATE_DEPARTMENT_SUCCESS = "CREATE_DEPARTMENT_SUCCESS",
  CREATE_DEPARTMENT_FAILURE = "CREATE_DEPARTMENT_FAILURE",
  UPDATE_DEPARTMENT_REQUEST = "UPDATE_DEPARTMENT_REQUEST",
  UPDATE_DEPARTMENT_SUCCESS = "UPDATE_DEPARTMENT_SUCCESS",
  UPDATE_DEPARTMENT_FAILURE = "UPDATE_DEPARTMENT_FAILURE",
  REMOVE_DEPARTMENT_REQUEST = "REMOVE_DEPARTMENT_REQUEST",
  REMOVE_DEPARTMENT_SUCCESS = "REMOVE_DEPARTMENT_SUCCESS",
  REMOVE_DEPARTMENT_FAILURE = "REMOVE_DEPARTMENT_FAILURE",
  SET_DEPARTMENT = "SET_DEPARTMENT",
  CLEAR_DEPARTMENT = "CLEAR_DEPARTMENT",
  SET_DEPARTMENTS = "SET_DEPARTMENTS",
  CLEAR_DEPARTMENTS = "CLEAR_DEPARTMENTS",
}